requirements.txt
requests==2.27.1
discord.py==1.7.3
humanize==4.0.0
